requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cloudinary==1.32.0
dj-database-url==2.0.0
Django==4.2.1
django-environ==0.10.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.5.0
urllib3==1.26.15
whitenoise==6.4.0
